Fix build issue

+++ src/common/include/com/amazonaws/kinesis/video/common/PlatformUtils.h
@@ -45,7 +45,7 @@ extern logPrintFunc globalCustomLogPrintFn;
 #ifdef DEBUG_BUILD
 #define __ASSERT(p1, p2, p3, ...) assert(p1)
 #else
-PUBLIC_API VOID customAssert(INT64 condition, const CHAR* fileName, int lineNumber, const CHAR* functionName);
+PUBLIC_API VOID customAssert(INT64 condition, const CHAR* fileName, INT64 lineNumber, const CHAR* functionName);
 #define __ASSERT(p1, p2, p3, ...) customAssert((p1), __FILE__, __LINE__, __FUNCTION__)
 #endif
